requirements.txt
python-binance>=1.0.16
python-dotenv
numpy
ta-lib
requests
pytest
xgboost==2.0.3
